Special case for event based
