+++ docs/effective-realizations.rst
@@ -12,7 +12,10 @@ full logic tree of a computation to a much smaller tree, containing
 much less effective realizations (i.e. paths) than the potential
 realizations.
 
-The reduction of the full logic tree happens when the actual
+Reduction of the GMPE logic tree
+------------------------------------
+
+The reduction of the GMPE logic tree happens when the actual
 sources do not span the full range of tectonic region types in the
 GMPE logic tree file. This happens practically always in SHARE calculations.
 The SHARE GMPE logic tree potentially contains 1280 realizations,
@@ -51,7 +54,7 @@ tectonic region type (T1) the GMPE logic tree file contains 3 GMPEs A,
 B, C and for the second tectonic region type (T2) the GMPE logic tree
 file contains 2 GMPEs D, E. The total number of realizations is
 
-total_num_rlzs = 3 * 2 = 6
+`total_num_rlzs = 3 * 2 = 6`
 
 The realizations are identified by an ordered pair of GMPEs, one for each
 tectonic region type. Let's number the realizations, starting from zero,
@@ -90,5 +93,40 @@ path of the effective realizations with the notation
 
 The engine lite will export two files with a name like
 
-hazard_curve-smltp_sm-gsimltp_*_D-ltr_0.csv
-hazard_curve-smltp_sm-gsimltp_*_E-ltr_1.csv
+`hazard_curve-smltp_sm-gsimltp_*_D-ltr_0.csv`, `hazard_curve-smltp_sm-gsimltp_*_E-ltr_1.csv`
+
+Reduction of the source model logic tree when sampling is enabled
+-----------------------------------------------------------------
+
+Consider a very common use case where one has a simple source model
+but a very large GMPE logic tree (we have real life examples
+with more than 400,000 branches). In such situation one would like to
+sample the branches of the GMPE logic tree. The complications is that
+currently the GMPE logic tree and the source model logic tree are
+coupled and the only way to sample the GMPE logic is to sample the
+source model logic tree. For each branch of the source model logic
+tree a single branch of the GMPE logic tree is chosen randomly,
+by taking into account the weights in the GMPE logic tree file.
+
+Suppose for instance that we set
+
+`number_of_logic_tree_samples = 4000`
+
+to sample 4,000 branches instead of 400,000. The expectation is
+that the computation will be 100 times faster, however this is
+not necessarily the case. There are two very different situations:
+
+1. if we are performing an event based calculation then each sample
+   of the source model will produce different ruptures even there is
+   only one source model repeated 4,000 times, because of the inherent
+   stochasticity of the process;
+2. if we are performing a classical (or disaggregation) calculation
+   identical samples will produce identical ruptures.
+
+In the second case it is possible to optimize the computation: if a
+source model path is sampled several times, we want to parse the
+model, send it to the workers and have it produce ruptures *only
+once*. In particular if there is a single source model and
+`number_of_logic_tree_samples = 4000`, we want to generate effectively
+1 source model realization and not 4,000 equivalent source model
+realizations.
