+++ openquake/qa_tests_data/classical/case_17/expected/hazard_curve-smltp_b2-gsimltp_b1-ltr_3.csv
@@ -1,0 +1,1 @@
+0.0 0.0,1.000000000E+00 1.000000000E+00 -0.000000000E+00
